gcp/requirements.txt
tensorflow-cpu==2.5.0
google-cloud-storage==2.8.0
pillow==8.0.0
